requirements.txt
Flask==3.0.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
yfinance==0.2.32
scipy==1.11.4
matplotlib==3.8.2
Pillow==10.1.0
Werkzeug==3.0.1
